swag-main/requirements.txt
Django~=4.2.7
django-phone-field~=1.8.1